+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Update after dropping deprecations from MP v2 - alloy
 - Add full screen carousel support in artwork view - ds300
 - Update after dropping duplicated types from MP v2 - zephraph
+- Updates textContentType for inputs on BillingAddress component - ash
 
 ### 1.12.11